requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
tensorflow==2.14.0
pillow==10.0.0
pdf2image==1.16.0